refactor: extract barrel runtime initialization into dedicated registry modules
