+++ src/services/index.ts
@@ -37,105 +37,8 @@ export * from './pilots';
 // Re-export game resolution services
 export * from './game-resolution';
 
-// ============================================================================
-// SERVICE REGISTRY
-// ============================================================================
-
-import {
-  getMechBuilderService,
-  getValidationService,
-  getCalculationService,
-} from './construction';
-import { getMTFImportService } from './conversion';
-import {
-  getEquipmentLoader,
-  getEquipmentRegistry,
-  getEquipmentNameMapper,
-} from './equipment';
-import { getIndexedDBService, getFileService } from './persistence';
-import { getPilotService, getPilotRepository } from './pilots';
-import { getUnitFactory } from './units';
-
-/**
- * Centralized service registry for singleton access
- * Uses factory functions for lazy initialization and better testability
- */
-export const services = {
-  // Persistence
-  persistence: {
-    get db() {
-      return getIndexedDBService();
-    },
-    get file() {
-      return getFileService();
-    },
-  },
-
-  // Equipment
-  equipment: {
-    get loader() {
-      return getEquipmentLoader();
-    },
-    get registry() {
-      return getEquipmentRegistry();
-    },
-    get nameMapper() {
-      return getEquipmentNameMapper();
-    },
-  },
-
-  // Units
-  units: {
-    get factory() {
-      return getUnitFactory();
-    },
-  },
-
-  // Construction
-  construction: {
-    get builder() {
-      return getMechBuilderService();
-    },
-    get validation() {
-      return getValidationService();
-    },
-    get calculation() {
-      return getCalculationService();
-    },
-  },
-
-  // Conversion
-  conversion: {
-    get importer() {
-      return getMTFImportService();
-    },
-  },
-
-  // Pilots
-  pilots: {
-    get service() {
-      return getPilotService();
-    },
-    get repository() {
-      return getPilotRepository();
-    },
-  },
-} as const;
-
-/**
- * Initialize all services that require async setup
- */
-export async function initializeServices(): Promise<void> {
-  // Initialize IndexedDB
-  await services.persistence.db.initialize();
-
-  // Initialize equipment registry
-  await services.equipment.registry.initialize();
-}
-
-/**
- * Shutdown all services
- */
-export function shutdownServices(): void {
-  services.persistence.db.close();
-}
+export {
+  initializeServices,
+  services,
+  shutdownServices,
+} from './serviceRegistry';
