+++ src/simulation/snapshot/snapshotMaintenance.ts
@@ -1,0 +1,6 @@
+import { SnapshotManager } from './SnapshotManager';
+
+export function cleanupOldSnapshots(days: number = 30): number {
+  const manager = new SnapshotManager();
+  return manager.deleteOldSnapshots(days);
+}
